+++ package.json
@@ -4,11 +4,11 @@
   "private": true,
   "dependencies": {
     "bootstrap": "^4.3.1",
-    "react": "^16.10.2",
     "react-bootstrap": "^1.0.0-beta.14",
+    "recharts": "^1.7.1",
+    "react": "^16.10.2",
     "react-dom": "^16.10.2",
-    "react-scripts": "3.2.0",
-    "recharts": "^1.7.1"
+    "react-scripts": "3.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
