+++ public/index.html
@@ -7,8 +7,12 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Coingraph is a real-time graph for cryptocurrencies."
+      content="Web site created using create-react-app"
     />
+    <meta
+    name="description"
+    content="Coingraph is a real-time graph for cryptocurrencies."
+  />
     <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
@@ -24,17 +28,14 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <title>Coingraph</title>
-    <!-- Bootstrap -->
     <link
-      rel="stylesheet"
-      href="node_modules/bootstrap/dist/css/bootstrap.min.css"
-    />
+    rel="stylesheet"
+    href="node_modules/bootstrap/dist/css/bootstrap.min.css"
+  />
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="btc"></div>
-    <div id="eth"></div>
-    
+    <div id="root"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
