+++ src/components/Bitcoin/index.test.js
