+++ src/components/Ethereum/index.test.js
